alfred/requirements.txt
pycontracts ~= 1.8.0
